Create a model using Keras
The ideal number of hidden layers and neurons depends on the problem and
the dataset. Like many aspects of machine learning, picking the best
shape of the neural network requires a mixture of knowledge and
experimentation.

As a rule of thumb, increasing the number of hidden layers and neurons
typically creates a more powerful model, which requires more data to
train effectively.